requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
cryptography==41.0.4
deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email-validator==2.0.0
flask==2.3.3
flask-limiter==3.5.0
flask-login==0.6.2
flask-mail==0.9.1
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
limits==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
openai==1.97.1
ordered-set==4.1.0
packaging==25.0
pip==24.2
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.0.0
rich==13.9.4
sniffio==1.3.1
sqlalchemy==2.0.21
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
werkzeug==2.3.7
wrapt==1.17.2
